+++ sdk/go/postgresql/config/config.go
@@ -10,12 +10,26 @@ import (
 
 // Maximum wait for connection, in seconds. Zero or not specified means wait indefinitely.
 func GetConnectTimeout(ctx *pulumi.Context) int {
-	return config.GetInt(ctx, "postgresql:connectTimeout")
+	v, err := config.TryInt(ctx, "postgresql:connectTimeout")
+	if err == nil {
+		return v
+	}
+	if dv, ok := getEnvOrDefault(0, parseEnvInt, "PGCONNECT_TIMEOUT").(int); ok {
+		return dv
+	}
+	return v
 }
 
 // The name of the database to connect to in order to conenct to (defaults to `postgres`).
 func GetDatabase(ctx *pulumi.Context) string {
-	return config.Get(ctx, "postgresql:database")
+	v, err := config.Try(ctx, "postgresql:database")
+	if err == nil {
+		return v
+	}
+	if dv, ok := getEnvOrDefault("", nil, "PGDATABASE").(string); ok {
+		return dv
+	}
+	return v
 }
 
 // Database username associated to the connected user (for user name maps)
@@ -30,7 +44,14 @@ func GetExpectedVersion(ctx *pulumi.Context) string {
 
 // Name of PostgreSQL server address to connect to
 func GetHost(ctx *pulumi.Context) string {
-	return config.Get(ctx, "postgresql:host")
+	v, err := config.Try(ctx, "postgresql:host")
+	if err == nil {
+		return v
+	}
+	if dv, ok := getEnvOrDefault("", nil, "PGHOST").(string); ok {
+		return dv
+	}
+	return v
 }
 
 // Maximum number of connections to establish to the database. Zero means unlimited.
@@ -40,12 +61,26 @@ func GetMaxConnections(ctx *pulumi.Context) int {
 
 // Password to be used if the PostgreSQL server demands password authentication
 func GetPassword(ctx *pulumi.Context) string {
-	return config.Get(ctx, "postgresql:password")
+	v, err := config.Try(ctx, "postgresql:password")
+	if err == nil {
+		return v
+	}
+	if dv, ok := getEnvOrDefault("", nil, "PGPASSWORD").(string); ok {
+		return dv
+	}
+	return v
 }
 
 // The PostgreSQL port number to connect to at the server host, or socket file name extension for Unix-domain connections
 func GetPort(ctx *pulumi.Context) int {
-	return config.GetInt(ctx, "postgresql:port")
+	v, err := config.TryInt(ctx, "postgresql:port")
+	if err == nil {
+		return v
+	}
+	if dv, ok := getEnvOrDefault(5432, parseEnvInt, "PGPORT").(int); ok {
+		return dv
+	}
+	return v
 }
 
 func GetSslMode(ctx *pulumi.Context) string {
@@ -55,7 +90,14 @@ func GetSslMode(ctx *pulumi.Context) string {
 // This option determines whether or with what priority a secure SSL TCP/IP connection will be negotiated with the
 // PostgreSQL server
 func GetSslmode(ctx *pulumi.Context) string {
-	return config.Get(ctx, "postgresql:sslmode")
+	v, err := config.Try(ctx, "postgresql:sslmode")
+	if err == nil {
+		return v
+	}
+	if dv, ok := getEnvOrDefault("", nil, "PGSSLMODE").(string); ok {
+		return dv
+	}
+	return v
 }
 
 // Specify if the user to connect as is a Postgres superuser or not.If not, some feature might be disabled (e.g.:
@@ -66,5 +108,12 @@ func GetSuperuser(ctx *pulumi.Context) bool {
 
 // PostgreSQL user name to connect as
 func GetUsername(ctx *pulumi.Context) string {
-	return config.Get(ctx, "postgresql:username")
+	v, err := config.Try(ctx, "postgresql:username")
+	if err == nil {
+		return v
+	}
+	if dv, ok := getEnvOrDefault("", nil, "PGUSER").(string); ok {
+		return dv
+	}
+	return v
 }