+++ sdk/python/pulumi_postgresql/config/vars.py
@@ -11,12 +11,12 @@ from .. import utilities, tables
 
 __config__ = pulumi.Config('postgresql')
 
-connect_timeout = __config__.get('connectTimeout')
+connect_timeout = __config__.get('connectTimeout') or utilities.get_env_int('PGCONNECT_TIMEOUT')
 """
 Maximum wait for connection, in seconds. Zero or not specified means wait indefinitely.
 """
 
-database = __config__.get('database')
+database = __config__.get('database') or utilities.get_env('PGDATABASE')
 """
 The name of the database to connect to in order to conenct to (defaults to `postgres`).
 """
@@ -31,7 +31,7 @@ expected_version = __config__.get('expectedVersion')
 Specify the expected version of PostgreSQL.
 """
 
-host = __config__.get('host')
+host = __config__.get('host') or utilities.get_env('PGHOST')
 """
 Name of PostgreSQL server address to connect to
 """
@@ -41,19 +41,19 @@ max_connections = __config__.get('maxConnections')
 Maximum number of connections to establish to the database. Zero means unlimited.
 """
 
-password = __config__.get('password')
+password = __config__.get('password') or utilities.get_env('PGPASSWORD')
 """
 Password to be used if the PostgreSQL server demands password authentication
 """
 
-port = __config__.get('port')
+port = __config__.get('port') or (utilities.get_env_int('PGPORT') or 5432)
 """
 The PostgreSQL port number to connect to at the server host, or socket file name extension for Unix-domain connections
 """
 
 ssl_mode = __config__.get('sslMode')
 
-sslmode = __config__.get('sslmode')
+sslmode = __config__.get('sslmode') or utilities.get_env('PGSSLMODE')
 """
 This option determines whether or with what priority a secure SSL TCP/IP connection will be negotiated with the
 PostgreSQL server
@@ -65,7 +65,7 @@ Specify if the user to connect as is a Postgres superuser or not.If not, some fe
 Refreshing state password from Postgres)
 """
 
-username = __config__.get('username')
+username = __config__.get('username') or utilities.get_env('PGUSER')
 """
 PostgreSQL user name to connect as
 """